backend/requirements.txt
Django>=6.0,<6.1
django-cors-headers>=4.3.0
djangorestframework>=3.14.0
gunicorn>=21.0.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9
